+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Holiday: St. Patrick's Day added.
 - PC sprites: Added Arrow Lizard, Koopa Troopa, Lolo, Lala, Imp, Magitek Officer, Locke, Gogo, Arus, Crono, Saul Goodman, and Madotsuki.
 - PC sprites: Improved vanilla Ness, Paula, Jeff, Poo, Mom, Dr Andonuts, and Shark.
+- Music: Custom track added.
 
 ## v50 (2022-02-11)
 - Ancient Cave: Jeff can fix any item at IQ 1.